backend/requirements.txt
flask
flask-cors
flask-pymongo
werkzeug
python-dotenv
PyJWT
